requirements.txt
pandas
numpy
faiss-cpu
langchain
langgraph
yfinance
newsapi-python
fredapi
google-generativeai
sec_api
diskcache
sentence-transformers
langchain-text-splitters
